Rename fileparsers module
